+++ src/react/src/components/ProductionLocation/Heading/LocationTitle/LocationTitle.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Typography from '@material-ui/core/Typography';
 import { withStyles } from '@material-ui/core/styles';
 
-import styles from './styles';
+import productionLocationDetailsTitleStyles from './styles';
 
 const ProductionLocationDetailsTitle = ({ classes }) => (
     <div className={classes.container}>
@@ -13,4 +13,6 @@ const ProductionLocationDetailsTitle = ({ classes }) => (
     </div>
 );
 
-export default withStyles(styles)(ProductionLocationDetailsTitle);
+export default withStyles(productionLocationDetailsTitleStyles)(
+    ProductionLocationDetailsTitle,
+);