+++ src/react/src/components/ProductionLocation/ProductionLocationDetailsContent/ProductionLocationDetailsContent.jsx
@@ -1,14 +1,8 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { Redirect, withRouter } from 'react-router';
-import { connect } from 'react-redux';
 import { withStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
 import Divider from '@material-ui/core/Divider';
-import CircularProgress from '@material-ui/core/CircularProgress';
-
-import isEmpty from 'lodash/isEmpty';
-import isArray from 'lodash/isArray';
 
 import ClaimFlag from '../Heading/ClaimFlag/ClaimFlag';
 import ClosureStatus from '../Heading/ClosureStatus/ClosureStatus';
@@ -19,193 +13,43 @@ import ClaimDataContainer from '../ClaimSection/ClaimDataContainer/ClaimDataCont
 import PartnerDataContainer from '../PartnerSection/PartnerDataContainer/PartnerDataContainer';
 import DetailsMap from '../ProductionLocationDetailsMap/ProductionLocationDetailsMap';
 
-import {
-    makeFacilityDetailLinkOnRedirect,
-    shouldUseProductionLocationPage,
-    getLastPathParameter,
-} from '../../../util/util';
+import productionLocationDetailsContentStyles from './styles';
 
-import { facilityClaimStatusChoicesEnum } from '../../../util/constants';
-
-import {
-    fetchSingleFacility,
-    resetSingleFacility,
-} from '../../../actions/facilities';
-
-import { facilityPropType } from '../../../util/propTypes';
-import styles from './styles';
-
-const ProductionLocationDetailsContent = ({
-    classes,
-    data,
-    fetching,
-    error,
-    contributors,
-    fetchFacility,
-    clearFacility,
-    location,
-    match: {
-        params: { osID },
-    },
-    useProductionLocationPage,
-    embed,
-}) => {
-    const normalizedOsID =
-        getLastPathParameter(location?.pathname || '') ||
-        getLastPathParameter(osID) ||
-        osID;
-
-    useEffect(() => {
-        fetchFacility(normalizedOsID, contributors);
-    }, [normalizedOsID, contributors]);
-
-    useEffect(() => () => clearFacility(), []);
-
-    if (fetching) {
-        return (
-            <div className={classes.root}>
-                <CircularProgress />
-            </div>
-        );
-    }
-
-    if (error && error.length) {
-        return (
-            <div className={classes.container}>
-                <ul>
-                    {error.map(err => (
-                        <li key={err} className={classes.error}>
-                            {err}
-                        </li>
-                    ))}
-                </ul>
-            </div>
-        );
-    }
-
-    if (!data) {
-        return (
-            <div className={classes.root}>
-                <p className={classes.primaryText}>
-                    {`No facility found for OS ID ${normalizedOsID}`}
-                </p>
-            </div>
-        );
-    }
-
-    if (data?.id && data?.id !== normalizedOsID) {
-        return (
-            <Redirect
-                to={makeFacilityDetailLinkOnRedirect(
-                    data.id,
-                    location.search,
-                    useProductionLocationPage,
-                )}
-            />
-        );
-    }
-
-    const isPendingClaim =
-        data?.properties?.claim_info?.status ===
-        facilityClaimStatusChoicesEnum.PENDING;
-    const isClaimed = !isPendingClaim && !!data?.properties?.claim_info;
-
-    return (
-        <div className={classes.container}>
-            <LocationTitle />
-            <ClaimFlag
-                osId={data.properties.os_id}
-                isClaimed={isClaimed}
-                isPending={isPendingClaim}
-                claimInfo={data?.properties?.claim_info}
-                isEmbed={!!embed}
-            />
-            <ClosureStatus
-                data={data}
-                clearFacility={clearFacility}
-                useProductionLocationPage={useProductionLocationPage}
-                search={location?.search || ''}
-            />
-            <DataSourcesInfo className={classes.containerItem} />
-            <Grid container xs={12} className={classes.containerItem}>
-                <Grid item sm={12} md={7}>
-                    <GeneralFields />
-                </Grid>
-                <Grid item sm={12} md={5}>
-                    <DetailsMap />
-                </Grid>
+const ProductionLocationDetailsContent = ({ classes }) => (
+    <div className={classes.container}>
+      <LocationTitle />
+          <ClaimFlag
+            osId={data.properties.os_id}
+            isClaimed={isClaimed}
+            isPending={isPendingClaim}
+            claimInfo={data?.properties?.claim_info}
+            isEmbed={!!embed}
+        />
+        <ClosureStatus
+            data={data}
+            clearFacility={clearFacility}
+            useProductionLocationPage={useProductionLocationPage}
+            search={location?.search || ''}
+        />
+        <DataSourcesInfo className={classes.containerItem} />
+        <Grid container className={classes.containerItem}>
+            <Grid item sm={12} md={7}>
+                <GeneralFields />
             </Grid>
-            <ClaimDataContainer className={classes.containerItem} />
-            <Divider variant="middle" className={classes.containerItem} />
-            <PartnerDataContainer />
-        </div>
-    );
-};
+            <Grid item sm={12} md={5}>
+                <DetailsMap />
+            </Grid>
+        </Grid>
+        <ClaimDataContainer className={classes.containerItem} />
+        <Divider variant="middle" className={classes.containerItem} />
+        <PartnerDataContainer />
+    </div>
+);
 
 ProductionLocationDetailsContent.propTypes = {
     classes: PropTypes.object.isRequired,
-    data: facilityPropType,
-    fetching: PropTypes.bool.isRequired,
-    error: PropTypes.arrayOf(PropTypes.string),
-    contributors: PropTypes.array,
-    fetchFacility: PropTypes.func.isRequired,
-    clearFacility: PropTypes.func.isRequired,
-    location: PropTypes.shape({
-        pathname: PropTypes.string,
-        search: PropTypes.string,
-    }).isRequired,
-    match: PropTypes.shape({
-        params: PropTypes.shape({
-            osID: PropTypes.string,
-        }).isRequired,
-    }).isRequired,
-    useProductionLocationPage: PropTypes.bool.isRequired,
-    embed: PropTypes.bool,
 };
 
-ProductionLocationDetailsContent.defaultProps = {
-    data: null,
-    error: [],
-    contributors: [],
-    embed: false,
-};
-
-function mapStateToProps({
-    facilities: {
-        singleFacility: { data, fetching, error },
-    },
-    filters: { contributors },
-    featureFlags,
-    embeddedMap: { embed } = {},
-}) {
-    return {
-        data,
-        fetching,
-        error,
-        contributors,
-        useProductionLocationPage: shouldUseProductionLocationPage(
-            featureFlags,
-        ),
-        embed: !!embed,
-    };
-}
-
-function mapDispatchToProps(dispatch) {
-    return {
-        fetchFacility: (id, contributorId) => {
-            const hasContributors =
-                isArray(contributorId) && !isEmpty(contributorId);
-            const contributors = hasContributors ? contributorId : null;
-
-            return dispatch(fetchSingleFacility(id, 0, contributors, true));
-        },
-        clearFacility: () => dispatch(resetSingleFacility()),
-    };
-}
-
-export default withRouter(
-    connect(
-        mapStateToProps,
-        mapDispatchToProps,
-    )(withStyles(styles)(ProductionLocationDetailsContent)),
+export default withStyles(productionLocationDetailsContentStyles)(
+    ProductionLocationDetailsContent,
 );