+++ src/react/src/components/ProductionLocation/ProductionLocationDetailsLocation/ProductionLocationDetailsLocation.jsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import { withStyles } from '@material-ui/core/styles';
 
-import styles from './styles';
+import productionLocationDetailsLocationStyles from './styles';
 
 const ProductionLocationDetailsLocation = () => (
     <>{/** TODO: refer to FacilityDetailsItem */}</>
 );
 
-export default withStyles(styles)(ProductionLocationDetailsLocation);
+export default withStyles(productionLocationDetailsLocationStyles)(
+    ProductionLocationDetailsLocation,
+);